chore: restructure folder and add webpack watch
