+++ src/client/Editor/components/Editor.tsx
@@ -1,8 +1,8 @@
 // Editor.tsx
 import React from 'react';
-import { Box, Button, TextField, MenuItem, Grid } from '@mui/material';
+import { Box, Button, TextField } from '@mui/material';
 import { useForm, Controller, useFieldArray } from 'react-hook-form';
-import { BusinessData, Section, Widget } from '../../types';
+import { BusinessData, Widget } from '../../../types';
 import WidgetRegistry from '../widgets/WidgetRegistry';
 
 interface EditorProps {
