+++ src/client/Editor/widgets/WidgetRegistry.tsx
