+++ src/client/styles/tailwind.scss
