+++ src/client/templates/mui/clientScript.tsx
