+++ src/client/templates/mui/components/AppBar.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from "react";
-import { BusinessData } from "../../types";
+import { BusinessData } from "../../../../types";
 import { AppBar as AppBarMui, Avatar, Box, Button, Container, IconButton, Menu, MenuItem, Toolbar, Tooltip, Typography } from '@mui/material';
 import AdbIcon from '@mui/icons-material/Adb';
 import MenuIcon from '@mui/icons-material/Menu';
