+++ src/client/templates/mui/components/Content.tsx
