+++ src/client/templates/mui/components/Section.tsx
